refactor(staking): move QBTC/Cosmos positions previews to PreviewActivity
Preview composables don't belong in the main source set — the @Preview
functions and their mock data compiled into the production APK. Per the
project convention (and #4806), screen previews live in the debug-only
PreviewActivity.

- Drop the two @Preview functions + preview host from
  CosmosStakingPositionsScreen; keep the stateless
  CosmosStakingPositionsContent extraction (now internal so debug can call it).
- Re-add the empty + one-active-delegation previews in PreviewActivity,
  wired into the screen dispatcher.
