+++ app/src/debug/java/com/vultisig/wallet/debug/PreviewActivity.kt
@@ -39,6 +39,7 @@ import androidx.compose.ui.unit.sp
 import com.vultisig.wallet.R
 import com.vultisig.wallet.data.blockchain.cosmos.qbtc.claim.QbtcClaimBlockedReason
 import com.vultisig.wallet.data.blockchain.cosmos.qbtc.claim.QbtcClaimError
+import com.vultisig.wallet.data.blockchain.cosmos.staking.CosmosStakePositionRow
 import com.vultisig.wallet.data.models.Address
 import com.vultisig.wallet.data.models.Chain
 import com.vultisig.wallet.data.models.Coins
@@ -70,6 +71,7 @@ import com.vultisig.wallet.ui.models.ChainTokensUiModel
 import com.vultisig.wallet.ui.models.TransactionDetailsUiModel
 import com.vultisig.wallet.ui.models.TransactionScanStatus
 import com.vultisig.wallet.ui.models.VerifyTransactionUiModel
+import com.vultisig.wallet.ui.models.cosmosstaking.CosmosStakingPositionsUiState
 import com.vultisig.wallet.ui.models.cosmosstaking.CosmosStakingVerifyUiState
 import com.vultisig.wallet.ui.models.cosmosstaking.CosmosStakingVerifyValidatorRow
 import com.vultisig.wallet.ui.models.deposit.DepositFormUiModel
@@ -92,6 +94,7 @@ import com.vultisig.wallet.ui.models.swap.ValuedToken
 import com.vultisig.wallet.ui.models.swap.VerifySwapUiModel
 import com.vultisig.wallet.ui.models.toNetworkUiModel
 import com.vultisig.wallet.ui.screens.TransactionDoneView
+import com.vultisig.wallet.ui.screens.cosmosstaking.CosmosStakingPositionsContent
 import com.vultisig.wallet.ui.screens.cosmosstaking.CosmosStakingVerifyContent
 import com.vultisig.wallet.ui.screens.deposit.BondFormContent
 import com.vultisig.wallet.ui.screens.keygen.FastVaultVerificationScreen
@@ -128,6 +131,7 @@ import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
 import dagger.hilt.android.EntryPointAccessors
 import dagger.hilt.components.SingletonComponent
+import java.math.BigDecimal
 
 @EntryPoint
 @InstallIn(SingletonComponent::class)
@@ -216,6 +220,9 @@ class PreviewActivity : ComponentActivity() {
                     "staking_verify_after" -> CosmosStakingVerifyCtaPreview(newButtons = true)
                     "staking_verify_qbtc" ->
                         CosmosStakingVerifyCtaPreview(newButtons = true, qbtc = true)
+                    "cosmos_staking_positions_empty" -> CosmosStakingPositionsEmptyPreview()
+                    "cosmos_staking_positions_delegation" ->
+                        CosmosStakingPositionsWithDelegationPreview()
                     "sign_message_before" -> VerifySignMessageCtaPreview(newButtons = false)
                     "sign_message_after" -> VerifySignMessageCtaPreview(newButtons = true)
                     else -> SwapConfirmPreview()
@@ -1911,4 +1918,75 @@ private fun ColumnScope.VerifySignMessageCta(newButtons: Boolean) {
             modifier = Modifier.fillMaxWidth(),
         )
     }
+}
+
+/**
+ * Empty QBTC staking state — reproduces issue #4831's screen (balance banner + "Staked" tab +
+ * "Delegate to New Validator" + empty-positions hint). The inline edit-chains icon that used to
+ * overlap the "Staked" label was removed in #4825, so this row now renders the tab alone.
+ */
+@Composable
+private fun CosmosStakingPositionsEmptyPreview() {
+    CosmosStakingPositionsPreviewHost(
+        state =
+            CosmosStakingPositionsUiState(
+                ticker = "QBTC",
+                selectedPositions = listOf("QBTC"),
+                totalAmountPrice = "$0.00",
+                totalStakedFiat = "$0.00",
+            )
+    )
+}
+
+/**
+ * Populated state — one active delegation with rewards, so the Claim button + a position card show.
+ */
+@Composable
+private fun CosmosStakingPositionsWithDelegationPreview() {
+    CosmosStakingPositionsPreviewHost(
+        state =
+            CosmosStakingPositionsUiState(
+                ticker = "QBTC",
+                selectedPositions = listOf("QBTC"),
+                totalStaked = BigDecimal("1250.5"),
+                totalAmountPrice = "$842.18",
+                totalStakedFiat = "$842.18",
+                hasClaimableRewards = true,
+                positions =
+                    listOf(
+                        CosmosStakePositionRow(
+                            validatorAddress =
+                                "terravaloper1qxv9z8m5n2k7w3pwl4rytsx0d8jh6c2e9abch7",
+                            validatorMoniker = "Vultisig Validator",
+                            validatorIdentity = null,
+                            stakedAmount = BigDecimal("1250.5"),
+                            stakedFiatDisplay = "$842.18",
+                            pendingReward = BigDecimal("0.453217"),
+                            apyPercent = BigDecimal("0.142"),
+                            validatorAvatarUrl = null,
+                            validatorStatus = CosmosStakePositionRow.ValidatorStatus.Active,
+                            pendingUnbondingUnlockDate = null,
+                            pendingUnbondingEntryCount = 0,
+                        )
+                    ),
+            )
+    )
+}
+
+@Composable
+private fun CosmosStakingPositionsPreviewHost(state: CosmosStakingPositionsUiState) {
+    CosmosStakingPositionsContent(
+        chainId = "QBTC",
+        state = state,
+        isRefreshing = false,
+        onRefresh = {},
+        onManagePositions = {},
+        onClaimAll = {},
+        onStakeMore = {},
+        onUnstake = {},
+        onMove = {},
+        onPositionSelectionChange = { _, _ -> },
+        onPositionSelectionDone = {},
+        onDismissPositionSelection = {},
+    )
 }
